refactor: Extract qos.rs from pipeline.rs
Move QoS statistics types (ElementQoSStats, QoSAggregator) to a
dedicated module. The methods using these types remain in mod.rs
but now import from the qos module.
